+++ LeFusion/dataset/__init__.py
@@ -2,3 +2,4 @@ from dataset.lidc_hist import LIDCDataset
 from dataset.lidc_hist_in import LIDCInDataset
 from dataset.emidec_hist import EMIDECDataset
 from dataset.emidec_hist_in import EMIDECInDataset
+from dataset.gli_hist import GLIDataset