+++ LeFusion/get_dataset/get_dataset.py
@@ -1,6 +1,7 @@
 from torch.utils.data import DataLoader
 from dataset import LIDCDataset, LIDCInDataset
 from dataset import EMIDECDataset, EMIDECInDataset
+from dataset import GLIDataset
 
 
 def get_inference_dataloader(dataset_root_dir, test_txt_dir,batch_size=1, drop_last=False, data_type=''):
@@ -23,6 +24,14 @@ def get_train_dataset(cfg):
     elif cfg.dataset.data_type == 'emidec':
         train_dataset = EMIDECDataset(root_dir=cfg.dataset.root_dir)
         sampler = None
+    elif cfg.dataset.data_type == 'gli':
+        train_dataset = GLIDataset(
+            root_dir=cfg.dataset.root_dir,
+            patch_size_xyz=cfg.dataset.patch_size_xyz,
+            split=cfg.dataset.get('split', 'train'),
+        )
+        sampler = None
+    else:
+        raise ValueError(f"Wrong data type: {cfg.dataset.data_type}")
     return train_dataset, sampler
-